moved cbmc to its own crate
