+++ compiler/cbmc/Cargo.toml
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 
 [package]
-name = "rustc_codegen_rmc"
+name = "cbmc"
 version = "0.0.0"
 edition = "2018"
 
@@ -12,7 +12,6 @@ doctest = false
 
 [dependencies]
 bitflags = "1.0"
-cbmc = { path = "../cbmc" }
 cstr = "0.2"
 libc = "0.2"
 measureme = "9.1.0"
