+++ compiler/cbmc/src/goto_program/symbol.rs
