+++ compiler/cbmc/src/goto_program/symbol_table.rs
