+++ compiler/cbmc/src/irep/symbol_table.rs
