+++ compiler/cbmc/src/utils.rs
