+++ compiler/rustc_codegen_rmc/src/mir_to_goto/codegen/assumptions.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 //! this module defines functions which impose data invariant on generated data types.
 
-use crate::cbmc::goto_program::{Expr, Location, Stmt, Symbol, Type};
 use crate::mir_to_goto::GotocCtx;
+use cbmc::goto_program::{Expr, Location, Stmt, Symbol, Type};
 use rustc_middle::mir::interpret::{ConstValue, Scalar};
 use rustc_middle::ty;
 use rustc_middle::ty::layout::LayoutOf;
