+++ compiler/rustc_codegen_rmc/src/mir_to_goto/codegen/mod.rs
@@ -15,3 +15,5 @@ mod span;
 mod statement;
 mod static_var;
 mod typ;
+
+pub use typ::TypeExt;
