+++ compiler/rustc_codegen_rmc/src/mir_to_goto/codegen/span.rs
@@ -3,8 +3,8 @@
 
 //! MIR Span related functions
 
-use crate::cbmc::goto_program::Location;
 use crate::mir_to_goto::GotocCtx;
+use cbmc::goto_program::Location;
 use rustc_middle::mir::{Local, VarDebugInfo, VarDebugInfoContents};
 use rustc_span::Span;
