+++ compiler/rustc_codegen_rmc/src/mir_to_goto/codegen/statement.rs
@@ -1,8 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+use super::typ::TypeExt;
 use super::typ::FN_RETURN_VOID_VAR_NAME;
-use crate::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Type};
 use crate::mir_to_goto::GotocCtx;
+use cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Type};
 use rustc_hir::def_id::DefId;
 use rustc_middle::mir;
 use rustc_middle::mir::{
