+++ compiler/rustc_codegen_rmc/src/mir_to_goto/overrides/hooks.rs
@@ -9,9 +9,9 @@
 //! this module addresses this issue.
 
 use super::stubs::{HashMapStub, VecStub};
-use crate::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Symbol, Type};
 use crate::mir_to_goto::utils::{instance_name_is, instance_name_starts_with};
 use crate::mir_to_goto::GotocCtx;
+use cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Symbol, Type};
 use rustc_hir::definitions::DefPathDataName;
 use rustc_middle::mir::{BasicBlock, Place};
 use rustc_middle::ty::layout::LayoutOf;
