+++ compiler/rustc_codegen_rmc/src/mir_to_goto/overrides/stubs/hash_map_stub.rs
@@ -10,8 +10,8 @@ use rustc_span::Span;
 use std::cell::{Cell, RefCell, RefMut};
 
 use super::super::hooks::GotocHook;
-use crate::cbmc::goto_program::{Expr, Location, Stmt};
 use crate::mir_to_goto::GotocCtx;
+use cbmc::goto_program::{Expr, Location, Stmt};
 
 pub struct HashMapStub<'tcx> {
     ty_opt: Cell<Option<Option<Ty<'tcx>>>>,
